spilt functionality over three classes
