+++ pom.xml
@@ -56,7 +56,7 @@
         <dependency>
             <groupId>com.weaverplatform</groupId>
             <artifactId>sdk-java</artifactId>
-            <version>1.2.2-SNAPSHOT</version>
+            <version>1.3.0-SNAPSHOT</version>
         </dependency>
         
         <!-- Testing -->
